mockgram firstcommit

+++ Mockgram/Mockgram-Server/mockgram-utils/utils/file-upload.js
@@ -15,28 +15,33 @@ checkImageType = (file) => {
 };
 
 exports.uploadImage = (options, file, cb) => {
-    if (checkImageType(file)) {
-        if (file.size > options.limit) {
-            return cb(response.ERROR.FILE_SIZE_EXCEEDED, null);
-        }
-        crypto.pseudoRandomBytes(16, (err, raw) => {
-            if (err) return cb(response.ERROR.SAVING_FILE_ERROR, null);
-            let fileName = Date.now() + "-" + raw.toString('hex') + path.extname(file.originalFilename);
-            let fileLocation = options.dest + fileName;
-            // connect-multiparty will creates temp files on server
-            // we should manually clean it after saving file to target path
-            let tmpPath = file.path;
-            fs.rename(tmpPath, fileLocation, (err) => {
+    if (file) {
+        if (checkImageType(file)) {
+            if (file.size > options.limit) {
+                return cb(response.ERROR.FILE_SIZE_EXCEEDED, null);
+            }
+            crypto.pseudoRandomBytes(16, (err, raw) => {
                 if (err) return cb(response.ERROR.SAVING_FILE_ERROR, null);
-                fs.unlink(tmpPath, () => {
+                let fileName = Date.now() + "-" + raw.toString('hex') + path.extname(file.originalFilename);
+                let fileLocation = options.dest + fileName;
+                // connect-multiparty will creates temp files on server
+                // we should manually clean it after saving file to target path
+                let tmpPath = file.path;
+                fs.rename(tmpPath, fileLocation, (err) => {
                     if (err) return cb(response.ERROR.SAVING_FILE_ERROR, null);
-                    // if everything ok, return the file name to the router 
-                    // data persistence needs file name
-                    return cb(null, fileName);
+                    fs.unlink(tmpPath, () => {
+                        if (err) return cb(response.ERROR.SAVING_FILE_ERROR, null);
+                        // if everything ok, return the file name to the router 
+                        // data persistence needs file name
+                        return cb(null, fileName);
+                    });
                 });
             });
-        });
+        } else {
+            return cb(response.ERROR.FILE_TYPE_ERROR, null);
+        }
+
     } else {
-        return cb(response.ERROR.FILE_TYPE_ERROR, null);
+        return cb(response.ERROR.NO_IMAGE_PROVIDED, null);
     }
 };
